rename modal listener, add create listener
